requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
streamlit
